+++ src/File Formats/BisUtils.P3D/Models/Lod/RVLodType.cs
