+++ src/File Formats/BisUtils.P3D/Models/Point/RVPoint.cs
@@ -4,6 +4,7 @@ using Core.Binarize.Options;
 using Core.Extensions;
 using Core.IO;
 using FResults;
+using Utils;
 
 public interface IRVPoint : IRVVector
 {
